pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scrape-store-agents"
version = "0.1.0"
description = "A simple, extensible web scraping and vector storage agent framework"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["web-scraping", "vector-database", "ai", "search", "agents"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "chromadb>=0.4.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "click>=8.1.0",
    "sentence-transformers>=2.2.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-openai>=0.0.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
scrape-store = "scrape_store_agents.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/scrape-store-agents"
Repository = "https://github.com/yourusername/scrape-store-agents.git"
Documentation = "https://github.com/yourusername/scrape-store-agents#readme"
"Bug Tracker" = "https://github.com/yourusername/scrape-store-agents/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["scrape_store_agents*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true